chapter-09/spot-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spot-python"
version = "1.0.0"
description = "Structured Prompt Output Toolkit - Python Implementation"
authors = [{name = "Chris Minnick", email = "chris@minnick.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
keywords = ["ai", "content-generation", "prompts", "templates"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "google-generativeai>=0.3.0",
    "httpx>=0.25.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.20.0",
    "PyYAML>=6.0.1",
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
]
dev = [
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]
cache = [
    "redis>=5.0.0",
]

[project.scripts]
spot = "spot.cli:main"

[project.urls]
Homepage = "https://github.com/chrisminnick/spot-toolkit"
Repository = "https://github.com/chrisminnick/spot-toolkit"
Documentation = "https://github.com/chrisminnick/spot-toolkit/tree/main/docs"

[tool.hatch.build.targets.wheel]
packages = ["spot"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--cov=spot --cov-report=term-missing --cov-report=html"